[Kernel] Fix PLE n-gram overflow and reference dispatch
Match PyTorch remainder semantics after signed int64 hash overflow, separate the PyTorch convolution reference path, and cover realistic mixed-batch token routing.

+++ vllm/models/qwen4_exp/nvidia/ops/ple_gate.py
@@ -76,7 +76,7 @@ def _ple_gate_kernel(
     if launch_pdl:
         tl.extra.cuda.gdc_launch_dependents()
     tl.store(gated_ptr + t * HC * H + offs, gated, mask=mask)
-    tl.store(normed_ptr + t * HC * H + offs, normed.to(tl.bfloat16), mask=mask)
+    tl.store(normed_ptr + t * HC * H + offs, normed, mask=mask)
 
 
 def _ple_gate(
